Fix bug with absolute file paths
Because `path.resolve` was being misused by passing a concatenated string directly rather than a series of path arguments, then passing an absolute path into the `env` function would not find the file in cases where it was loaded relative to a parent directory (it concatenated the path to itself).

This should have manifested in a failing test, but the test itself was flawed so that it used an environment variable left over by a previous test. Resolve that issue by explicitly removing the particular environment variable under test at the beginning of the test.